add supplemental panel
